Deploying-ML-Models/requirements.txt
#requirements
numpy == 1.18.1
pandas == 1.0.3
Django == 2.2.10
djangorestframework == 3.9.1
gunicorn == 20.0.4
#psycopg2 == 2.8.4
#django-heroku == 0.3.1
sklearn == 0.0
tensorflow == 1.15.0
Werkzeug==1.0.1
imutils == 0.5.3
opencv-python == 4.1.0.25
dlib ==19.17.0
scipy == 1.4.1
google-cloud-storage == 1.10.0